Taking matters into my own hands
The form is now the dialog body.

The dialog itself is just the backdrop (so I can detect it being clicked)

I removed default styling for the dialog and the backdrop

It works swell. Focus is even disabled inside the dialog as it animates closed. Excellent! I am very happy with this

But does it work in other browsers?